requirements.txt
requests
beautifulsoup4
PyYAML
PyMuPDF
python-docx
python-pptx
pandas
pydub
opencv-python
moviepy
ffmpeg-python
torch
torchaudio
librosa
Pillow
openai-whisper
fastapi
uvicorn
rich
python-multipart
# 其它依赖可根据实际业务补充
